Add files via upload
A computational project using MATLAB in which I developed a MATLAB programme to solve problems in linear thermo-elasticity, with linear triangular elements. This resulted in authoring a ten page report using LaTeX, of which a few select pages are showcased here. Both this project and the corresponding report were graded, achieving a combined 2:1.